Move full text search to proper events
