+++ lib/Event/AclCreatedEvent.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+
+namespace OCA\Deck\Event;
+
+class AclCreatedEvent extends AAclEvent {
+}
